feat(cmd-004): Stage D — requester-route ui_intent server-side over WS
The spec's Key mechanics: a command's UI intent is rendered by the surface
that ISSUED the command — not broadcast. Pre-Stage-D the WS handler forwarded
every ui_intent to every connected client (proven RED: 3 failing routing
tests). Now `subscribeSessionEvents` delivers an intent only when
`event.requesterDriverId` matches THIS surface's server-assigned driver id,
read lazily (`getSurfaceDriverId`) so the SessionResumeBridge's late-bound
pairing id (setDriverId) routes correctly. An unattributed intent (idle
model-invoked command) is unroutable and reaches every surface — never a
silent drop. In the bridge, routing precedes seq-stamping/buffering, so a
foreign surface's intent consumes no seq and cannot leak through a `resume`
replay.

The outbound session→TServerMessage fan-out is split into
`ws-session-events.ts` (file-size ratchet: ws-handler.ts stays under its
frozen baseline; inbound routing stays in ws-handler.ts).
