+++ packages/agent-transport-protocol/src/ws-handler.ts
@@ -12,21 +12,15 @@ import {
   handleBackgroundControlMessage,
   handleBackgroundQueryMessage,
 } from './ws-background-messages.js';
+import { subscribeSessionEvents } from './ws-session-events.js';
 
 import type { TClientMessage, TServerMessage } from './ws-protocol.js';
-import type { TDriverId } from '@robota-sdk/agent-interface-transport';
-import type {
-  IAskRequestEvent,
-  IExecutionResult,
-  IExecutionWorkspaceEvent,
-  IInteractiveSession,
-  IPermissionRequestEvent,
-  IPromptResolvedEvent,
-  IToolState,
-  IUiIntentEvent,
-  TBackgroundJobGroupEvent,
-} from '@robota-sdk/agent-interface-transport';
-import type { TBackgroundTaskEvent } from '@robota-sdk/agent-interface-transport';
+import type { IInteractiveSession, TDriverId } from '@robota-sdk/agent-interface-transport';
+
+// Outbound session→TServerMessage fan-out (incl. CMD-004 requester-routed `ui_intent`) lives in
+// `ws-session-events.ts`; re-exported here for the bridge and existing importers.
+export { subscribeSessionEvents } from './ws-session-events.js';
+export type { ISubscribeSessionEventsOptions } from './ws-session-events.js';
 
 export interface IWsHandlerOptions {
   /** IInteractiveSession to expose. */
@@ -63,89 +57,12 @@ export function createWsHandler(options: IWsHandlerOptions): {
   onMessage: (data: string) => void;
   cleanup: () => void;
 } {
-  const cleanup = subscribeSessionEvents(options.session, options.send);
+  const cleanup = subscribeSessionEvents(options.session, options.send, {
+    getSurfaceDriverId: () => options.driverId,
+  });
   const onMessage = createWsMessageHandler(options.session, options.send, options.driverId);
 
   return { onMessage, cleanup };
-}
-
-/**
- * Subscribe the session's events and forward each as a `TServerMessage` via `send`; returns an unsubscribe.
- * Exported (REMOTE-013 E4) so the persistent {@link SessionResumeBridge} can own a SINGLE subscription that
- * outlives per-channel handlers.
- */
-export function subscribeSessionEvents(
-  session: IInteractiveSession,
-  send: (message: TServerMessage) => void,
-): () => void {
-  // REMOTE-014 E5: stamp the ACTIVE turn's driver id onto TURN-AUTHORED events (co-drive authorship,
-  // display-only), read at emit time. Only these events — background/goal/memory/execution-workspace events
-  // are NOT authored by a driver turn and carry no `driverId`. `undefined` when idle or unattributed.
-  const attr = (): { driverId?: TDriverId } => {
-    const d = session.getActiveDriverId?.() ?? undefined;
-    return d ? { driverId: d } : {};
-  };
-  const onUserMessage = (content: string): void =>
-    send({ type: 'user_message', content, ...attr() });
-  const onTextDelta = (delta: string): void => send({ type: 'text_delta', delta, ...attr() });
-  const onToolStart = (state: IToolState): void => send({ type: 'tool_start', state, ...attr() });
-  const onToolEnd = (state: IToolState): void => send({ type: 'tool_end', state, ...attr() });
-  const onThinking = (isThinking: boolean): void =>
-    send({ type: 'thinking', isThinking, ...attr() });
-  const onComplete = (result: IExecutionResult): void =>
-    send({ type: 'complete', result, ...attr() });
-  const onInterrupted = (result: IExecutionResult): void =>
-    send({ type: 'interrupted', result, ...attr() });
-  const onError = (error: Error): void =>
-    send({ type: 'error', message: error.message, ...attr() });
-  const onBackgroundTaskEvent = (event: TBackgroundTaskEvent): void =>
-    send({ type: 'background_task_event', event });
-  const onBackgroundJobGroupEvent = (event: TBackgroundJobGroupEvent): void =>
-    send({ type: 'background_job_group_event', event });
-  const onExecutionWorkspace = (event: IExecutionWorkspaceEvent): void =>
-    send({ type: 'execution_workspace_event', snapshot: event.snapshot });
-  // REMOTE-007: forward the transport-neutral prompt events so a remote surface can render + answer the
-  // SAME permission/ask prompt; `prompt_resolved` dismisses it when another surface answered first.
-  const onPermissionRequest = (event: IPermissionRequestEvent): void =>
-    send({ type: 'permission_request', event });
-  const onAskRequest = (event: IAskRequestEvent): void => send({ type: 'ask_request', event });
-  const onPromptResolved = (event: IPromptResolvedEvent): void =>
-    send({ type: 'prompt_resolved', event });
-  const onUiIntent = (event: IUiIntentEvent): void => send({ type: 'ui_intent', event }); // CMD-004
-
-  session.on('user_message', onUserMessage);
-  session.on('text_delta', onTextDelta);
-  session.on('tool_start', onToolStart);
-  session.on('tool_end', onToolEnd);
-  session.on('thinking', onThinking);
-  session.on('complete', onComplete);
-  session.on('interrupted', onInterrupted);
-  session.on('error', onError);
-  session.on('background_task_event', onBackgroundTaskEvent);
-  session.on('background_job_group_event', onBackgroundJobGroupEvent);
-  session.on('execution_workspace_event', onExecutionWorkspace);
-  session.on('permission_request', onPermissionRequest);
-  session.on('ask_request', onAskRequest);
-  session.on('prompt_resolved', onPromptResolved);
-  session.on('ui_intent', onUiIntent);
-
-  return (): void => {
-    session.off('user_message', onUserMessage);
-    session.off('text_delta', onTextDelta);
-    session.off('tool_start', onToolStart);
-    session.off('tool_end', onToolEnd);
-    session.off('thinking', onThinking);
-    session.off('complete', onComplete);
-    session.off('interrupted', onInterrupted);
-    session.off('error', onError);
-    session.off('background_task_event', onBackgroundTaskEvent);
-    session.off('background_job_group_event', onBackgroundJobGroupEvent);
-    session.off('execution_workspace_event', onExecutionWorkspace);
-    session.off('permission_request', onPermissionRequest);
-    session.off('ask_request', onAskRequest);
-    session.off('prompt_resolved', onPromptResolved);
-    session.off('ui_intent', onUiIntent);
-  };
 }
 
 function createWsMessageHandler(
@@ -222,11 +139,7 @@ function isSessionQueryMessage(msg: TClientMessage): msg is Extract<
   TClientMessage,
   {
     type:
-      | 'get-messages'
-      | 'get-context'
-      | 'get-executing'
-      | 'get-pending'
-      | 'get-execution-workspace';
+      'get-messages' | 'get-context' | 'get-executing' | 'get-pending' | 'get-execution-workspace';
   }
 > {
   return (
